requirements.txt
django
django-bootstrap4
django-crispy-forms
dj-database-url
dj-static
djangorestframework
python-decouple
psycopg2-binary
gunicorn
whitenoise
Unipath
djongo